Move towards recognizing Roman numerals properly
Now it's at least capable of matching 3+-character Roman numerals
(e.g., "viii"), and once it picks up that it's looking at Roman
numerals, subsequent prefixes will be assumed to be Roman numerals
unless proven otherwise.

What this doesn't solve is the initial match. "i" will be identified as
a letter, and "ii" will likewise be identified as letters. Not until
"iii" will the Roman numeral be matched. But it's a start.